customer pay order

+++ ecommerce-order-service/src/main/kotlin/me/luraframework/example/order/command/PayOrderCommand.kt
@@ -3,5 +3,5 @@ package me.luraframework.example.order.command
 import java.math.BigDecimal
 
 data class PayOrderCommand(
-  val payPrice: Int
+  val payPrice: BigDecimal
 )
